search for day by the day prop, not the order of the array

+++ src/problems/aocday.rs
@@ -1,5 +1,5 @@
 pub struct AoCDay {
-	pub day: u8,
+	pub day: usize,
 	pub part1: Box<dyn Fn(&std::path::PathBuf)>,
 	pub part2: Box<dyn Fn(&std::path::PathBuf)>
 }
